backend/requirements.txt
Django>=5.0,<6.0
django-ninja>=1.3.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
django-cors-headers>=4.0.0
gunicorn>=21.0.0
python-decouple>=3.8
openpyxl>=3.1.0
